Initial Commit
Three part program progression.

Ps1c uses binary search to find the best savings rate to make a down payment on a home in 36 months

Finding the right amount to save away
In Part B, you had a chance to explore how both the percentage of your salary that you save each month
and your annual raise affect how long it takes you to save for a down payment.  This is nice, but
suppose you want to set a particular goal, e.g. to be able to afford the down payment in three years.
How much should you save each month to achieve this?  In this problem, you are going to write a
program to answer that question.  To simplify things, assume:
3
1. Your semi­annual raise is .07 (7%)
2. Your investments have an annual return of 0.04 (4%)
3. The down payment is 0.25 (25%) of the cost of the house
4. The cost of the house that you are saving for is $1M.

Enter the starting salary: 150000
Best savings rate: 0.4411